Improve Message Blocks
