+++ src/common/util/copyToClipboard.ts
@@ -1,0 +1,6 @@
+export function copyToClipboard(text: string) {
+  if (typeof navigator !== 'undefined')
+    navigator.clipboard.writeText(text)
+      .then(() => console.log('Message copied to clipboard'))
+      .catch((err) => console.error('Failed to copy message: ', err));
+}
